+++ src/renderer/world_renderer.hpp
@@ -24,8 +24,8 @@ public:
 
 public: // r_IWorldRenderer
 	virtual void Update() override;
-	virtual void UpdateChunk( unsigned short X, unsigned short Y, bool immediately ) override;
-	virtual void UpdateChunkWater( unsigned short X, unsigned short Y, bool immediately ) override;
+	virtual void UpdateChunk( unsigned int X, unsigned int Y, bool immediately ) override;
+	virtual void UpdateChunkWater( unsigned int X, unsigned int Y, bool immediately ) override;
 	virtual void UpdateWorldPosition( int longitude, int latitude ) override;
 
 public:
